[FEATURE & DESIGN] 마이페이지 및 통계페이지 1차 작업

+++ src/screens/MyPage.js
@@ -12,6 +12,7 @@ import {
 } from 'react-native';
 import Footer from '../components/footer';
 import authClient from '../apis/authClient';
+import Lottie from 'lottie-react-native';
 
 const MyPage = ({navigation}) => {
   const [selectedMenu, setSelectedMenu] = useState(null);
@@ -121,27 +122,47 @@ const MyPage = ({navigation}) => {
 
   // 회원탈퇴
   const handleWithdraw = async () => {
-    try {
-      const res = await authClient({
-        method: 'delete',
-        url: '/users',
-      });
-      // LandingPage로 navigate
-      navigation.navigate('LandingPage');
-      await AsyncStorage.clear();
-      Alert.alert('회원탈퇴 되었습니다.');
-    } catch (error) {
-      console.log('회원탈퇴 실패', error);
-    }
+    // 회원탈퇴 다이얼로그 추가
+    Alert.alert(
+      '회원 탈퇴 확인',
+      '정말로 회원을 탈퇴하시겠습니까?',
+      [
+        {
+          text: '예',
+          onPress: async () => {
+            try {
+              const res = await authClient({
+                method: 'delete',
+                url: '/users',
+              });
+              // LandingPage로 navigate
+              navigation.navigate('LandingPage');
+              await AsyncStorage.clear();
+              Alert.alert('회원 탈퇴되었습니다.');
+            } catch (error) {
+              console.log(error);
+            }
+          },
+        },
+        {
+          text: '아니요',
+          style: 'cancel',
+        },
+      ],
+      { cancelable: false }
+    );
   };
+
 
   return (
     <View style={styles.container}>
       <View style={styles.profileContainer}>
-        <Image
-          source={require('../../assets/gold.png')}
-          style={styles.avatar}
-        />
+      <Lottie
+      source={require('../../assets/animations/newBear.json')}
+      style={styles.image}
+      autoPlay
+      loop
+    />
         <Text style={styles.username}>{user ? user.username : '이름'}</Text>
         <Text style={styles.age}>
           {user ? `나이: ${user.age}살` : '나이: ?'}
@@ -325,6 +346,13 @@ const styles = StyleSheet.create({
   confirmButtonText: {
     color: 'white',
     fontSize: 18,
+  },
+  image: {
+    width: 100, // 이미지의 너비 설정
+    height: 100, // 이미지의 높이 설정
+    borderRadius: 75, // 반지름을 절반으로 설정하여 동그랗게 만듭니다.
+    backgroundColor: 'black',
+    marginBottom: 10,
   },
 });
